requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
django-ckeditor==5.8.0
django-simple-mail==2.3.1
django-import-export==2.0.1
djangorestframework==3.11.0
djoser==2.0.3